Refactor: 디렉토리 위치, 대소문자 변경
