+++ hustle_web/src/components/Community/Lecture/LectureStyle.ts
