+++ hustle_web/src/components/Competition/List1.tsx
