+++ hustle_web/src/components/Competition/Styles.ts
