+++ hustle_web/src/components/CompetitionList/Styles.tsx
