+++ hustle_web/src/components/Content/Styles.tsx
