+++ hustle_web/src/pages/Login/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {LoginContainer} from "./styles";
+import {LoginContainer} from "./Styles";
 
 
 const Login = () => {
